extract dropdown selector and interactor methods
